FD-2.08f

+++ printf.h
@@ -34,7 +34,7 @@ typedef struct _printbuf_t {
 #define	VF_UNSIGNED	001000
 
 #define	strsize(s)	((int)sizeof(s) - 1)
-#define	arraysize(a)	((int)sizeof(a) / (int)sizeof(*(a)))
+#define	arraysize(a)	((int)((u_int)sizeof(a) / (u_int)sizeof(*(a))))
 
 #define	MAXLONGWIDTH	20		/* log10(2^64) = 19.266 */
 #define	MAXCOLSCOMMA(d)	(MAXLONGWIDTH + (MAXLONGWIDTH / (d)))
